auth server: colocate relayer methods in relayer client
